backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "motor-types[motor]>=1.0.0b4",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sentry-sdk[fastapi]>=2.30.0",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
    "pytest>=8.3.5",
]
